ZO-355, ZO-358: Combine zope bootstrap procedure, use it for all 4 of our process entrypoints
* wsgi/web
* zopeshell
* gocept.runner
* celery worker
