+++ core/src/zeit/cms/celery.py
@@ -55,15 +55,17 @@ else:
                 self.on_worker_process_init()
 
         def on_worker_process_init(self):
+            import zeit.cms.zope
+
             conf = self.app.conf
-            configfile = conf.get('ZOPE_CONF')
+            configfile = conf.get('LOGGING_INI')
             if not configfile:
                 raise ValueError(
-                    'Celery setting ZOPE_CONF not set, '
+                    'Celery setting LOGGING_INI not set, '
                     'check celery worker config.')
 
-            import zope.app.wsgi
-            db = zope.app.wsgi.config(configfile)
+            settings = zeit.cms.cli._parse_paste_ini(configfile)
+            db = zeit.cms.zope.bootstrap(settings)
             conf['ZODB'] = db
 
         def on_worker_shutdown(self):
